# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ski Area Dashboard API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: lifts
    description: Lift passage operations
  - name: dashboard
    description: Dashboard summary operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /lifts/latest:
    get:
      operationId: getLatestLifts
      tags: [lifts]
      summary: Get latest snapshot for all lifts
      description: Returns the most recent extraction snapshot per lift for the given date. If extraction is provided, pins to that specific extraction time.
      parameters:
        - name: season
          in: query
          required: false
          schema:
            type: string
          description: Season filter e.g. "2024-2025"
        - name: date
          in: query
          required: false
          schema:
            type: string
          description: Date filter in YYYY-MM-DD format, defaults to today
        - name: extraction
          in: query
          required: false
          schema:
            type: string
          description: Specific extraction time (dupd value e.g. '20211219230000'). If omitted returns latest per lift.
      responses:
        "200":
          description: List of latest lift snapshots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LiftSnapshot"

  /lifts/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary totals
      description: Returns aggregated totals for the dashboard
      parameters:
        - name: season
          in: query
          required: false
          schema:
            type: string
        - name: date
          in: query
          required: false
          schema:
            type: string
        - name: extraction
          in: query
          required: false
          schema:
            type: string
          description: Pin to a specific extraction time (dupd value)
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /lifts/dates:
    get:
      operationId: getLiftDates
      tags: [lifts]
      summary: List all dates that have lift data
      description: Returns distinct dates (YYYY-MM-DD) ordered most recent first
      parameters:
        - name: season
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of date strings in YYYY-MM-DD format, most recent first
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

  /lifts/extractions:
    get:
      operationId: getLiftExtractions
      tags: [lifts]
      summary: List available extraction times for a date
      description: Returns distinct dupd values (extraction timestamps) for a given date, ordered ascending
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
          description: Date in YYYY-MM-DD format, defaults to today
        - name: season
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of extraction time strings (dupd values)
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

  /lifts/history:
    get:
      operationId: getLiftHistory
      tags: [lifts]
      summary: Get historical snapshots for a specific lift
      parameters:
        - name: ggnr
          in: query
          required: true
          schema:
            type: integer
        - name: season
          in: query
          required: false
          schema:
            type: string
        - name: date
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of snapshots for this lift on the given date
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LiftSnapshot"
        "404":
          description: Lift not found

  /lifts/sync:
    post:
      operationId: syncLifts
      tags: [lifts]
      summary: Push extraction data from on-site script
      description: Receives a batch of lift snapshots from the scheduled extraction script
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncInput"
      responses:
        "200":
          description: Sync result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncResult"
        "401":
          description: Unauthorized

  /lifts/period:
    get:
      operationId: getLiftsPeriod
      tags: [lifts]
      summary: Get aggregated lift stats for a date range
      description: Returns per-lift cumulative totals and an overall summary for the given date range. Each lift's daily latest snapshot is summed across the period.
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
          description: Start date in YYYY-MM-DD format
        - name: to
          in: query
          required: true
          schema:
            type: string
          description: End date in YYYY-MM-DD format (inclusive)
        - name: season
          in: query
          required: false
          schema:
            type: string
          description: Season filter e.g. "2024-2025"
      responses:
        "200":
          description: Period aggregated data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PeriodResult"
        "400":
          description: Invalid parameters

  /seasons:
    get:
      operationId: getSeasons
      tags: [lifts]
      summary: List available seasons
      responses:
        "200":
          description: List of season strings
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

  /lifts/season-trend:
    get:
      operationId: getSeasonTrend
      tags: [lifts]
      summary: Get daily passage totals per season for comparison charts
      description: Returns daily aggregated totals per season so the client can draw multi-season comparison lines. Each season entry contains an array of daily data points ordered by date.
      parameters:
        - name: seasons
          in: query
          required: false
          schema:
            type: string
          description: Comma-separated list of season strings e.g. "2023-2024,2024-2025". Defaults to the two most recent seasons.
      responses:
        "200":
          description: Season trend data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SeasonTrend"

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    LiftSnapshot:
      type: object
      required: [id, ggnr, ggbz, nsoc, dtgg, dupd, eser]
      properties:
        id:
          type: integer
        idin:
          type: ["string", "null"]
        dupd:
          type: string
          description: Date and time of extraction
        dtgg:
          type: string
          description: Date of the data (ISO 8601)
        ggnr:
          type: integer
          description: Lift code
        ggbz:
          type: string
          description: Lift description/name
        nsoc:
          type: ["integer", "null"]
          description: Lift sort/group number
        npin:
          type: ["integer", "null"]
          description: Guests with first passage of the day
        npic:
          type: ["integer", "null"]
          description: Guests with second passage in different company
        nuin:
          type: ["integer", "null"]
          description: Number of guests on lifts at extraction time
        npas:
          type: ["integer", "null"]
          description: Total number of passages
        eser:
          type: string
          description: Season (e.g. "2024-2025")
        nomeSocieta:
          type: ["string", "null"]
          description: Company name
        descrGrp:
          type: ["string", "null"]
          description: Company group description
        idSocieta:
          type: ["integer", "null"]
          description: Company ID
        codgrp:
          type: ["string", "null"]
          description: Company group code

    DashboardSummary:
      type: object
      required: [date, season, totalPassages, totalGuests, totalLifts, activeLifts, lastSyncAt]
      properties:
        date:
          type: string
        season:
          type: string
        totalPassages:
          type: integer
          description: Sum of npas across all lifts
        totalGuests:
          type: integer
          description: Sum of nuin across all lifts
        totalLifts:
          type: integer
          description: Total number of lifts
        activeLifts:
          type: integer
          description: Lifts with at least one passage
        lastSyncAt:
          type: ["string", "null"]
          description: ISO timestamp of last data sync

    SyncInput:
      type: object
      required: [snapshots]
      properties:
        snapshots:
          type: array
          items:
            $ref: "#/components/schemas/SnapshotInput"

    SnapshotInput:
      type: object
      required: [dupd, dtgg, ggnr, ggbz, eser]
      properties:
        idin:
          type: ["string", "null"]
        dupd:
          type: string
        dtgg:
          type: string
        ggnr:
          type: integer
        ggbz:
          type: string
        nsoc:
          type: ["integer", "null"]
        npin:
          type: ["integer", "null"]
        npic:
          type: ["integer", "null"]
        nuin:
          type: ["integer", "null"]
        npas:
          type: ["integer", "null"]
        eser:
          type: string
        nome_societa:
          type: ["string", "null"]
          description: Company name
        descr_grp:
          type: ["string", "null"]
          description: Company group description
        id_societa:
          type: ["string", "null"]
          description: Company ID
        codgrp:
          type: ["string", "null"]
          description: Company group code

    SyncResult:
      type: object
      required: [inserted, updated, total]
      properties:
        inserted:
          type: integer
        updated:
          type: integer
        total:
          type: integer
        message:
          type: ["string", "null"]

    PeriodLiftRow:
      type: object
      required: [ggnr, ggbz, totalPassages, totalGuests, totalFirstPassages, activeDays]
      properties:
        ggnr:
          type: integer
          description: Lift code
        ggbz:
          type: string
          description: Lift name
        totalPassages:
          type: integer
          description: Sum of daily-latest npas across the period
        totalGuests:
          type: integer
          description: Sum of daily-latest nuin across the period
        totalFirstPassages:
          type: integer
          description: Sum of daily-latest npin across the period
        activeDays:
          type: integer
          description: Number of days the lift had at least one passage

    SeasonTrendPoint:
      type: object
      required: [date, dayIndex, totalPassages, totalGuests]
      properties:
        date:
          type: string
          description: Date in YYYY-MM-DD format
        dayIndex:
          type: integer
          description: 1-based index of this day within the season (for aligning seasons on the x-axis)
        totalPassages:
          type: integer
          description: Total passages across all lifts on this date
        totalGuests:
          type: integer
          description: Total guests on lifts across all lifts on this date

    SeasonTrend:
      type: object
      required: [season, data]
      properties:
        season:
          type: string
          description: Season identifier e.g. "2024-2025"
        data:
          type: array
          items:
            $ref: "#/components/schemas/SeasonTrendPoint"
          description: Daily data points ordered by date

    PeriodResult:
      type: object
      required: [from, to, totalPassages, totalGuests, activeDays, lifts]
      properties:
        from:
          type: string
          description: Start date YYYY-MM-DD
        to:
          type: string
          description: End date YYYY-MM-DD
        season:
          type: ["string", "null"]
        totalPassages:
          type: integer
          description: Sum of all lift passages in the period
        totalGuests:
          type: integer
          description: Sum of all guest counts in the period
        activeDays:
          type: integer
          description: Number of distinct days that had any data
        busiestDay:
          type: ["string", "null"]
          description: YYYY-MM-DD date with the most passages
        busiestLift:
          type: ["string", "null"]
          description: Name of the lift with the most passages
        lifts:
          type: array
          items:
            $ref: "#/components/schemas/PeriodLiftRow"
          description: Per-lift aggregated totals, sorted by total passages descending