# lib/api-zod/src/generated/types/seasonTrend.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Ski Area Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { SeasonTrendPoint } from './seasonTrendPoint';

export interface SeasonTrend {
  /** Season identifier e.g. "2024-2025" */
  season: string;
  /** Daily data points ordered by date */
  data: SeasonTrendPoint[];
}