# lib/api-zod/src/generated/types/seasonTrendPoint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Ski Area Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface SeasonTrendPoint {
  /** Date in YYYY-MM-DD format */
  date: string;
  /** 1-based index of this day within the season (for aligning seasons on the x-axis) */
  dayIndex: number;
  /** Total passages across all lifts on this date */
  totalPassages: number;
  /** Total guests on lifts across all lifts on this date */
  totalGuests: number;
}